collector/requirements.txt
psutil>=5.9.0
requests>=2.28.0
schedule>=1.1.0
PyYAML>=6.0
python-dotenv>=0.20.0
